Break project into gradle modules
