+++ buildSrc/build.gradle.kts
@@ -1,0 +1,22 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    // Support convention plugins written in Kotlin. Convention plugins are build scripts in 'src/main' that automatically become available as plugins in the main build.
+    `kotlin-dsl`
+}
+
+repositories {
+    // Use the plugin portal to apply community plugins in convention plugins.
+    gradlePluginPortal()
+}
+
+dependencies {
+    implementation(libs.kotlin.gradle.plugin)
+    implementation(plugin(libs.plugins.kotlin.plugin.serialization))
+}
+
+
+fun plugin(plugin: Provider<PluginDependency>) = plugin.get().run { plugin(pluginId, version.requiredVersion) }
+fun plugin(id: String, version: String) = "$id:$id.gradle.plugin:$version"
