+++ buildSrc/src/main/kotlin/buildlogic.kotlin-library-conventions.gradle.kts
@@ -1,0 +1,11 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    // Apply the common convention plugin for shared build configuration between library and application projects.
+    id("buildlogic.kotlin-common-conventions")
+
+    // Apply the java-library plugin for API and implementation separation.
+    `java-library`
+}
