+++ example-domain/application/build.gradle.kts
@@ -1,0 +1,38 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    id("buildlogic.kotlin-library-conventions")
+    id("org.jetbrains.kotlin.plugin.serialization")
+}
+
+tasks.named<Jar>("jar") { archiveFileName.set("example-domain-application.jar") }
+
+dependencies {
+    api(project(":example-domain:domain"))
+    api(project(":example-domain:infrastructure"))
+    api(project(":example-domain:core"))
+
+    // ## ktor API ##
+    implementation(libs.ktor.server.core.jvm)
+    implementation(libs.ktor.serialization.kotlinx.json.jvm)
+    implementation(libs.ktor.server.content.negotiation.jvm)
+    implementation(libs.ktor.server.request.validation)
+    implementation(libs.ktor.server.host.common.jvm)
+    implementation(libs.ktor.server.status.pages.jvm)
+    implementation(libs.ktor.server.netty.jvm)
+
+    // ## Functional Programming ##
+    implementation(libs.arrow.core)
+
+    // ## Database ##
+    implementation(libs.jooq.kotlin)
+
+    // ## Logging ##
+    implementation(libs.logback.classic)
+
+    // ## Testing ##
+    testImplementation(libs.ktor.server.tests.jvm)
+    testImplementation(libs.kotlin.test.junit)
+}
