+++ example-domain/application/src/main/kotlin/com/fsn/template/application/account/adapter/request/UpdateAccountApiRequest.kt
