+++ example-domain/application/src/main/kotlin/com/fsn/template/application/account/controller/AccountsController.kt
