+++ example-domain/application/src/main/kotlin/com/fsn/template/application/configuration/Serialization.kt
