+++ example-domain/application/src/main/resources/application.conf
