+++ example-domain/application/src/main/resources/logback.xml
