+++ example-domain/core/build.gradle.kts
@@ -1,0 +1,17 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    id("buildlogic.kotlin-library-conventions")
+}
+
+tasks.named<Jar>("jar") { archiveFileName.set("example-domain-core.jar") }
+
+dependencies {
+    // ## Functional Programming ##
+    implementation(libs.arrow.core)
+
+    // ## Testing ##
+    testImplementation(libs.kotlin.test.junit)
+}
