+++ example-domain/core/src/main/kotlin/com/fsn/template/core/ClockHelper.kt
