+++ example-domain/domain/build.gradle.kts
@@ -1,0 +1,19 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    id("buildlogic.kotlin-library-conventions")
+}
+
+tasks.named<Jar>("jar") { archiveFileName.set("example-domain-domain.jar") }
+
+dependencies {
+    implementation(project(":example-domain:core"))
+
+    // ## Functional Programming ##
+    implementation(libs.arrow.core)
+
+    // ## Testing ##
+    testImplementation(libs.kotlin.test.junit)
+}
