+++ example-domain/domain/src/main/kotlin/com/fsn/template/domain/account/command/UpdateAccountCommand.kt
