+++ example-domain/infrastructure/src/main/kotlin/com/fsn/template/infrastructure/account/SqlAccountRepository.kt
