+++ example-domain/infrastructure/src/main/resources/db/migration/V2__Create_accounts_table.sql
