--- gradle.properties
@@ -1,9 +1,0 @@
-kotlin.code.style=official
-ktor_version=2.3.12
-kotlin_version=2.0.0
-logback_version=1.4.14
-flyway_version=9.20.1
-hikari_version=5.1.0
-mysql_version=8.0.33
-jooq_kotlin_version=3.19.10
-arrow_version=1.2.4
