+++ gradle/libs.versions.toml
@@ -1,0 +1,53 @@
+# This file was generated by the Gradle 'init' task.
+# https://docs.gradle.org/current/userguide/platforms.html#sub::toml-dependencies-format
+
+[versions]
+kotlin-gradle-plugin = "2.0.0"
+kotlin.code.style = "official"
+ktor_version = "2.3.12"
+kotlin_version = "2.0.0"
+logback_version = "1.4.14"
+flyway_version = "9.20.1"
+hikari_version = "5.1.0"
+mysql_version = "8.0.33"
+jooq_kotlin_version = "3.19.10"
+arrow_version = "1.2.4"
+
+[libraries]
+# Plugin
+kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", version.ref = "kotlin-gradle-plugin" }
+jooq-codegen-gradle-plugin = { module = "org.jooq:jooq-meta-extensions", version.ref = "jooq_kotlin_version" }
+
+## ktor API ##
+ktor-server-core-jvm = { module = "io.ktor:ktor-server-core-jvm", version.ref = "ktor_version" }
+ktor-serialization-kotlinx-json-jvm = { module = "io.ktor:ktor-serialization-kotlinx-json-jvm", version.ref = "ktor_version" }
+ktor-server-content-negotiation-jvm = { module = "io.ktor:ktor-server-content-negotiation-jvm", version.ref = "ktor_version" }
+ktor-server-request-validation = { module = "io.ktor:ktor-server-request-validation", version.ref = "ktor_version" }
+ktor-server-host-common-jvm = { module = "io.ktor:ktor-server-host-common-jvm", version.ref = "ktor_version" }
+ktor-server-status-pages-jvm = { module = "io.ktor:ktor-server-status-pages-jvm", version.ref = "ktor_version" }
+ktor-server-netty-jvm = { module = "io.ktor:ktor-server-netty-jvm", version.ref = "ktor_version" }
+
+## Functional Programming ##
+arrow-core = { module = "io.arrow-kt:arrow-core", version.ref = "arrow_version" }
+
+## DB ##
+jooq-kotlin = { module = "org.jooq:jooq-kotlin", version.ref = "jooq_kotlin_version" }
+jooq-kotlin-coroutines = { module = "org.jooq:jooq-kotlin-coroutines", version.ref = "jooq_kotlin_version" }
+
+# Required to work with datetime in DB
+hikari-cp = { module = "com.zaxxer:HikariCP", version.ref = "hikari_version" }
+mysql-connector-java = { module = "mysql:mysql-connector-java", version.ref = "mysql_version" }
+
+## Flyway ##
+flyway-core = { module = "org.flywaydb:flyway-core", version.ref = "flyway_version" }
+flyway-mysql = { module = "org.flywaydb:flyway-mysql", version.ref = "flyway_version" }
+
+## Logging ##
+logback-classic = { module = "ch.qos.logback:logback-classic", version.ref = "logback_version" }
+
+## Testing ##
+ktor-server-tests-jvm = { module = "io.ktor:ktor-server-tests-jvm", version.ref = "ktor_version" }
+kotlin-test-junit = { module = "org.jetbrains.kotlin:kotlin-test-junit", version.ref = "kotlin_version" }
+
+[plugins]
+kotlin-plugin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin_version" }
