--- src/test/kotlin/com/fsn/template/application/ApplicationTest.kt
@@ -1,20 +1,0 @@
-package com.fsn.template.application
-
-import com.fsn.template.application.configuration.configureHealth
-import io.ktor.client.request.get
-import io.ktor.client.statement.bodyAsText
-import io.ktor.http.HttpStatusCode
-import io.ktor.server.testing.testApplication
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ApplicationTest {
-  @Test
-  fun testRoot() = testApplication {
-    application { configureHealth() }
-    client.get("/health").apply {
-      assertEquals(HttpStatusCode.OK, status)
-      assertEquals("ok", bodyAsText())
-    }
-  }
-}
