--- src/test/resources/application.conf
@@ -1,8 +1,0 @@
-ktor {
-    deployment {
-        port = 8080
-    }
-    application {
-        modules = [ com.fsn.template.application.ApplicationKt.testModule ]
-    }
-}
